+++ crates/jarl-core/src/lints/lengths/lengths.rs
@@ -1,5 +1,5 @@
 use crate::diagnostic::*;
-use crate::utils::{get_arg_by_name_then_position, node_contains_comments};
+use crate::utils::{get_arg_by_name_then_position, get_function_name, node_contains_comments};
 use air_r_syntax::*;
 use anyhow::Context;
 use biome_rowan::AstNode;
@@ -48,9 +48,10 @@ impl Violation for Lengths {
 pub fn lengths(ast: &RCall) -> anyhow::Result<Option<Diagnostic>> {
     let RCallFields { function, arguments } = ast.as_fields();
     let function = function?;
+    let fn_name = get_function_name(function);
 
     let funs_to_watch = ["sapply", "vapply", "map_dbl", "map_int"];
-    if !funs_to_watch.contains(&function.into_syntax().text_trimmed().to_string().as_str()) {
+    if !funs_to_watch.contains(&fn_name.as_str()) {
         return Ok(None);
     }