+++ crates/jarl-core/src/lints/system_file/system_file.rs
@@ -1,5 +1,5 @@
 use crate::diagnostic::*;
-use crate::utils::{get_named_args, get_unnamed_args, node_contains_comments};
+use crate::utils::{get_function_name, get_named_args, get_unnamed_args, node_contains_comments};
 use air_r_syntax::*;
 use biome_rowan::{AstNode, AstSeparatedList};
 pub struct SystemFile;
@@ -43,9 +43,10 @@ impl Violation for SystemFile {
 
 pub fn system_file(ast: &RCall) -> anyhow::Result<Option<Diagnostic>> {
     let function = ast.function()?;
+    let fn_name = get_function_name(function);
     let arguments = ast.arguments()?;
 
-    if function.to_trimmed_text() != "system.file" {
+    if fn_name != "system.file" {
         return Ok(None);
     }