+++ docs/changelog.md
@@ -5,6 +5,8 @@
 ### Features
 
 - `--output-format json` now contains two fields `diagnostics` and `errors` (#219).
+- Better support for namespaced function calls, both when reporting violations
+  and when fixing them (#221).
 
 ## 0.2.1